+++ openrouteservice/src/main/java/heigit/ors/api/requests/routing/RouteRequestHandler.java
@@ -23,10 +23,26 @@ import heigit.ors.api.requests.common.APIEnums;
 import heigit.ors.api.requests.common.GenericHandler;
 import heigit.ors.common.DistanceUnit;
 import heigit.ors.common.StatusCode;
-import heigit.ors.exceptions.*;
+import heigit.ors.exceptions.EmptyElementException;
+import heigit.ors.exceptions.IncompatibleParameterException;
+import heigit.ors.exceptions.InternalServerException;
+import heigit.ors.exceptions.ParameterOutOfRangeException;
+import heigit.ors.exceptions.ParameterValueException;
+import heigit.ors.exceptions.StatusCodeException;
+import heigit.ors.exceptions.UnknownParameterValueException;
 import heigit.ors.geojson.GeometryJSON;
 import heigit.ors.localization.LocalizationManager;
-import heigit.ors.routing.*;
+import heigit.ors.routing.AvoidFeatureFlags;
+import heigit.ors.routing.RouteExtraInfoFlag;
+import heigit.ors.routing.RouteInstructionsFormat;
+import heigit.ors.routing.RouteResult;
+import heigit.ors.routing.RouteSearchParameters;
+import heigit.ors.routing.RoutingErrorCodes;
+import heigit.ors.routing.RoutingProfileManager;
+import heigit.ors.routing.RoutingProfileType;
+import heigit.ors.routing.RoutingRequest;
+import heigit.ors.routing.WayPointBearing;
+import heigit.ors.routing.WeightingMethod;
 import heigit.ors.routing.graphhopper.extensions.reader.borders.CountryBordersReader;
 import heigit.ors.routing.pathprocessors.BordersExtractor;
 import heigit.ors.util.DistanceUnitUtil;
@@ -80,9 +96,14 @@ public class RouteRequestHandler extends GenericHandler {
         if (request.hasAttributes())
             routingRequest.setAttributes(convertAttributes(request.getAttributes()));
 
-        if (request.hasExtraInfo())
+        if (request.hasExtraInfo()){
             routingRequest.setExtraInfo(convertExtraInfo(request.getExtraInfo()));
-
+            for (APIEnums.ExtraInfo extra: request.getExtraInfo()) {
+                if (extra.compareTo(APIEnums.ExtraInfo.COUNTRY_INFO) == 0) {
+                    routingRequest.setIncludeCountryInfo(true);
+                }
+            }
+        }
         if (request.hasLanguage())
             routingRequest.setLanguage(convertLanguage(request.getLanguage()));