+++ openrouteservice/src/main/java/heigit/ors/routing/graphhopper/extensions/storages/builders/BordersGraphStorageBuilder.java
@@ -1,15 +1,15 @@
 /*  This file is part of Openrouteservice.
  *
- *  Openrouteservice is free software; you can redistribute it and/or modify it under the terms of the 
- *  GNU Lesser General Public License as published by the Free Software Foundation; either version 2.1 
+ *  Openrouteservice is free software; you can redistribute it and/or modify it under the terms of the
+ *  GNU Lesser General Public License as published by the Free Software Foundation; either version 2.1
  *  of the License, or (at your option) any later version.
 
- *  This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
- *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ *  This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  *  See the GNU Lesser General Public License for more details.
 
- *  You should have received a copy of the GNU Lesser General Public License along with this library; 
- *  if not, see <https://www.gnu.org/licenses/>.  
+ *  You should have received a copy of the GNU Lesser General Public License along with this library;
+ *  if not, see <https://www.gnu.org/licenses/>.
  */
 package heigit.ors.routing.graphhopper.extensions.storages.builders;
 
@@ -20,7 +20,6 @@ import com.graphhopper.util.EdgeIteratorState;
 import com.vividsolutions.jts.geom.Coordinate;
 import com.vividsolutions.jts.geom.GeometryFactory;
 import com.vividsolutions.jts.geom.LineString;
-import com.vividsolutions.jts.geom.Point;
 import heigit.ors.exceptions.MissingConfigParameterException;
 import heigit.ors.routing.graphhopper.extensions.reader.borders.CountryBordersPolygon;
 import heigit.ors.routing.graphhopper.extensions.reader.borders.CountryBordersReader;
@@ -43,6 +42,8 @@ public class BordersGraphStorageBuilder extends AbstractGraphStorageBuilder {
     private CountryBordersReader cbReader;
 
     private GeometryFactory gf;
+
+    public static String builderName = "Borders";
 
     public BordersGraphStorageBuilder() {
         gf = new GeometryFactory();
@@ -119,13 +120,15 @@ public class BordersGraphStorageBuilder extends AbstractGraphStorageBuilder {
     public void processWay(ReaderWay way, Coordinate[] coords, HashMap<Integer, HashMap<String,String>> nodeTags) {
         // Process the way using the geometry provided
         // if we don't have the reader object, then we can't do anything
-        if(cbReader != null) {
+        if (cbReader != null) {
             String[] countries = findBorderCrossing(coords);
-
             // If we find that the length of countries is more than one, then it does cross a border
             if (countries.length > 1 && !countries[0].equals(countries[1])) {
                 way.setTag("country1", countries[0]);
                 way.setTag("country2", countries[1]);
+            } else if (countries.length == 1){
+                way.setTag("country1", countries[0]);
+                way.setTag("country2", countries[0]);
             }
         }
     }
@@ -139,35 +142,30 @@ public class BordersGraphStorageBuilder extends AbstractGraphStorageBuilder {
      * @param way  The OSM way obtained from the OSM reader. This way corresponds to the edge to be processed
      * @param edge The graph edge to be process
      */
-    @Override
-    public void processEdge(ReaderWay way, EdgeIteratorState edge) {
-        // Make sure we actually have the storage initialised - if there were errors accessing the data then this could be the case
-        if(_storage != null) {
-            // If there is no border crossing then we set the edge value to be 0
+     @Override
+     public void processEdge(ReaderWay way, EdgeIteratorState edge) {
+         // Make sure we actually have the storage initialised - if there were errors accessing the data then this could be the case
+         if (_storage != null) {
+             // If there is no border crossing then we set the edge value to be 0
 
-            // First get the start and end countries - if either of these is empty, then there is no crossing
-            if (way.hasTag("country1") && way.hasTag("country2")) {
-                String startVal = way.getTag("country1");
-                String endVal = way.getTag("country2");
-
-                // Lookup values
-                short start = 0, end = 0;
-
-                try {
-                    start = Short.parseShort(cbReader.getId(startVal));
-                    end = Short.parseShort(cbReader.getId(endVal));
-                } catch (NumberFormatException nfe) {
-                    LOGGER.error("Error in lookup for ids " + startVal + " and " + endVal);
-                }
-
-                short type = (cbReader.isOpen(cbReader.getEngName(startVal), cbReader.getEngName(endVal))) ? (short) 2 : (short) 1;
-
-                _storage.setEdgeValue(edge.getEdge(), type, start, end);
-            } else {
-                _storage.setEdgeValue(edge.getEdge(), (short) 0, (short) 0, (short) 0);
-            }
-        }
-    }
+             // First get the start and end countries - if they are equal, then there is no crossing
+             String startVal = way.getTag("country1");
+             String endVal = way.getTag("country2");
+             short type = BordersGraphStorage.NO_BORDER;
+             short start = 0;
+             short end = 0;
+             try {
+                 start = Short.parseShort(cbReader.getId(startVal));
+                 end = Short.parseShort(cbReader.getId(endVal));
+             } catch (Exception ignore) {
+             } finally {
+                 if (start != end) {
+                     type = (cbReader.isOpen(cbReader.getEngName(startVal), cbReader.getEngName(endVal))) ? (short) 2 : (short) 1;
+                 }
+                 _storage.setEdgeValue(edge.getEdge(), type, start, end);
+             }
+         }
+     }
 
     /**
      * Method identifying the name of the extension which is used in various building processes
@@ -176,7 +174,7 @@ public class BordersGraphStorageBuilder extends AbstractGraphStorageBuilder {
      */
     @Override
     public String getName() {
-        return "Borders";
+        return builderName;
     }
 
     /**
@@ -257,6 +255,7 @@ public class BordersGraphStorageBuilder extends AbstractGraphStorageBuilder {
             // Replace the arraylist
             countries = temp;
         }
+
         // Now we have a list of all the countries that the nodes are in - if this is more than one it is likely it is
         // crossing a border, but not certain as in some disputed areas, countries overlap and so it may not cross any
         // border.
@@ -296,5 +295,9 @@ public class BordersGraphStorageBuilder extends AbstractGraphStorageBuilder {
         }
 
         return names.toArray(new String[names.size()]);
+    }
+
+    public CountryBordersReader getCbReader() {
+        return cbReader;
     }
 }